make validation optional and fix bug with chromadb http client (#8221)
* make validation optional and fix bug with chromadb http client

* clearer

* add api_key as accepted input

* move logic out of base class init

* refactor chroma parameter validation
fix chroma tests not running sequentially

* remove from base class
