+++ mindsdb/integrations/handlers/chromadb_handler/chromadb_handler.py
@@ -3,6 +3,7 @@ from typing import List, Optional
 
 import pandas as pd
 
+from mindsdb.integrations.handlers.chromadb_handler.settings import ChromaHandlerConfig
 from mindsdb.integrations.libs.const import HANDLER_CONNECTION_ARG_TYPE as ARG_TYPE
 from mindsdb.integrations.libs.response import RESPONSE_TYPE
 from mindsdb.integrations.libs.response import HandlerResponse
@@ -14,6 +15,7 @@ from mindsdb.integrations.libs.vectordatabase_handler import (
     TableField,
     VectorStoreHandler,
 )
+from mindsdb.interfaces.storage.model_fs import HandlerStorage
 from mindsdb.utilities import log
 
 
@@ -48,16 +50,39 @@ class ChromaDBHandler(VectorStoreHandler):
     name = "chromadb"
 
     def __init__(self, name: str, **kwargs):
-        super().__init__(name, **kwargs)
+        super().__init__(name)
+        self.handler_storage = HandlerStorage(kwargs.get("integration_id"))
+        self._client = None
+        self.persist_directory = None
+        self.is_connected = False
+
+        config = self.validate_connection_parameters(name, **kwargs)
 
         self._client_config = {
-            "chroma_server_host": self.config.host,
-            "chroma_server_http_port": self.config.port,
+            "chroma_server_host": config.host,
+            "chroma_server_http_port": config.port,
             "persist_directory": self.persist_directory,
         }
 
-        self._client = None
         self.connect()
+
+    def validate_connection_parameters(self, name, **kwargs):
+        """
+        Validate the connection parameters.
+        """
+
+        _config = kwargs.get("connection_data")
+        _config["vector_store"] = name
+
+        config = ChromaHandlerConfig(**_config)
+
+        if config.persist_directory and not self.handler_storage.is_temporal:
+            # get full persistence directory from handler storage
+            self.persist_directory = self.handler_storage.folder_get(
+                config.persist_directory
+            )
+
+        return config
 
     def _get_client(self):
         client_config = self._client_config
@@ -76,7 +101,14 @@ class ChromaDBHandler(VectorStoreHandler):
             )
 
     def __del__(self):
-        super().__del__()
+        """Close the database connection."""
+
+        if self.is_connected is True:
+            if self.persist_directory:
+                # sync folder to handler storage
+                self.handler_storage.folder_sync(self.persist_directory)
+
+            self.disconnect()
 
     def connect(self):
         """Connect to a ChromaDB database."""
@@ -88,8 +120,8 @@ class ChromaDBHandler(VectorStoreHandler):
             self.is_connected = True
             return self._client
         except Exception as e:
-            log.logger.error(f"Error connecting to ChromaDB client, {e}!")
             self.is_connected = False
+            raise Exception(f"Error connecting to ChromaDB client, {e}!")
 
     def disconnect(self):
         """Close the database connection."""
@@ -377,7 +409,7 @@ connection_args = OrderedDict(
         "required": False,
     },
     port={
-        "type": ARG_TYPE.INT,
+        "type": ARG_TYPE.STR,
         "description": "chromadb server port",
         "required": False,
     },
@@ -390,6 +422,6 @@ connection_args = OrderedDict(
 
 connection_args_example = OrderedDict(
     host="localhost",
-    port=8000,
+    port="8000",
     persist_directory="chroma",
 )
