+++ mindsdb/integrations/libs/vectordatabase_handler.py
@@ -1,5 +1,4 @@
 import ast
-import difflib
 import uuid
 from enum import Enum
 from typing import Any, List, Optional
@@ -18,10 +17,8 @@ from mindsdb_sql.parser.ast import (
     Update,
 )
 from mindsdb_sql.parser.ast.base import ASTNode
-from pydantic import BaseModel, Extra, root_validator
 
 from mindsdb.integrations.libs.response import RESPONSE_TYPE, HandlerResponse
-from mindsdb.interfaces.storage.model_fs import HandlerStorage
 from mindsdb.utilities.log import get_log
 
 from ..utilities.sql_utils import query_traversal
@@ -94,79 +91,6 @@ class TableField(Enum):
     DISTANCE = "distance"
 
 
-class VectorStoreHandlerConfig(BaseModel):
-    """
-    Configuration for VectorStoreHandler.
-    """
-
-    vector_store: str
-    persist_directory: str = None
-    host: str = None
-    port: int = None
-    url: str = None
-
-    class Config:
-        extra = Extra.forbid
-
-    @root_validator(pre=True, allow_reuse=True)
-    def check_param_typos(cls, values):
-        """Check if there are any typos in the parameters."""
-
-        expected_params = cls.__fields__.keys()
-        for key in values.keys():
-            if key not in expected_params:
-                close_matches = difflib.get_close_matches(
-                    key, expected_params, cutoff=0.4
-                )
-                if close_matches:
-                    raise ValueError(
-                        f"Unexpected parameter '{key}'. Did you mean '{close_matches[0]}'?"
-                    )
-                else:
-                    raise ValueError(f"Unexpected parameter '{key}'.")
-        return values
-
-    @root_validator(allow_reuse=True, skip_on_failure=True)
-    def check_config(cls, values):
-        """Check if config is valid."""
-
-        vector_store = values.get("vector_store")
-        host = values.get("host")
-        port = values.get("port")
-        url = values.get("url")
-        persist_directory = values.get("persist_directory")
-
-        if host and not port:
-            raise ValueError(
-                f"For {vector_store} handler - if host is provided, port must also be provided."
-            )
-
-        if port and not host:
-            raise ValueError(
-                f"For {vector_store} handler - if port is provided, host must also be provided."
-            )
-
-        if port and host and (url or persist_directory):
-            raise ValueError(
-                f"For {vector_store} handler - if host and port are provided, "
-                f"url and persistence_folder should not be provided."
-            )
-
-        if url and (host or port or persist_directory):
-            raise ValueError(
-                f"For {vector_store} handler - if url is provided, host, port, "
-                f"persist_directory should not be provided."
-            )
-
-        if persist_directory and (url or host or port):
-            raise ValueError(
-                f"For {vector_store} handler - if persistence_folder is provided, "
-                f"url, host, port should not be provided."
-            )
-
-        return values
-
-
 class VectorStoreHandler(BaseHandler):
     """
     Base class for handlers associated to vector databases.
@@ -191,34 +115,13 @@ class VectorStoreHandler(BaseHandler):
         },
     ]
 
-    def __init__(self, name: str, **kwargs):
+    def __init__(self, name: str):
         super().__init__(name)
-        kwargs["connection_data"].pop("password", None)
-        self.handler_storage = HandlerStorage(kwargs.get("integration_id"))
 
-        _config = kwargs.get("connection_data")
-        _config["vector_store"] = name
+    def validate_connection_parameters(self, name, **kwargs):
+        """Create validation for input parameters."""
 
-        self.config = VectorStoreHandlerConfig(**_config)
-
-        self.is_connected = False
-
-        if self.config.persist_directory and not self.handler_storage.is_temporal:
-            # get full persistence directory from handler storage
-            self.persist_directory = self.handler_storage.folder_get(
-                self.config.persist_directory
-            )
-
-    def __del__(self):
-        if self.is_connected is True:
-            if self.persist_directory:
-                # sync folder to handler storage
-                self.handler_storage.folder_sync(self.persist_directory)
-
-            self.disconnect()
-
-    def disconnect(self):
-        raise NotImplementedError()
+        return NotImplementedError()
 
     def _value_or_self(self, value):
         if isinstance(value, Constant):
